+++ R/prepare_elements.R
@@ -82,6 +82,11 @@ prepare_elements_from_BED12 = function(fname) {
   coords$start = as.numeric(coords$start)
   coords$end = as.numeric(coords$end)
   # coords$frag_id = as.numeric(coords$frag_id) to exclude frag_id as a necessary column
+  if (any(coords$start < 2) | any(coords$end < 2)) {
+    bad_ids = unique(coords$id[coords$start < 2 | coords$end < 2])
+    stop(paste0("elements contain coordinates below 2 for the following ids: ",
+        paste(bad_ids, collapse = ", ")))
+  }
   dim1 = nrow(coords)
 
   # Filtering Regions for Those Identified on a Single Chromosome (Mitochondrial, Autosomal and Sex Chromosomes)
@@ -141,6 +146,11 @@ prepare_elements_from_BED4 = function(fname) {
   input = data.frame(input, stringsAsFactors = F)
   input$start = as.numeric(input$start)
   input$end = as.numeric(input$end)
+  if (any(input$start < 2) | any(input$end < 2)) {
+    bad_ids = unique(input$id[input$start < 2 | input$end < 2])
+    stop(paste0("elements contain coordinates below 2 for the following ids: ",
+        paste(bad_ids, collapse = ", ")))
+  }
   dim1 = nrow(input)
   
   # Filtering Regions for Those Identified on a Single Chromosome (Mitochondrial, Autosomal and Sex Chromosomes)